LayoutTests/security/decode-buffer-size-expected.txt

ALERT: successfully caught exception RangeError: Bad value
ALERT: successfully caught exception RangeError: Bad value
